requirements.txt
click==6.7
pkg-resources==0.0.0
python-vlc==3.0.101
